Fix mixed indents on cftemplates (#50)
* Fix mixed indents on snapshots_tool_rds_dest.josn

snapshots_tool_rds_dest.json has a mix of spaces and tabs used for indentation. This converts the only instance of spaces to Tabs to ensure consistency.

* Fix mixed indents on snapshots_tool_rds_source

snapshots_tool_rds_source.json has a mix of spaces and tabs used for indentation. This converts the only instance of spaces to Tabs to ensure consistency.